+++ R/double_variable.R
@@ -21,12 +21,15 @@ DoubleVariable <- R6Class(
     get_values = function(index = NULL) {
       if (is.null(index)) {
         return(double_variable_get_values(self$.variable))
+      } else {
+        if (inherits(index, 'Bitset')) {
+          return(double_variable_get_values_at_index(self$.variable, index$.bitset))
+        } else {
+          stopifnot(all(is.finite(index)))
+          stopifnot(all(index > 0))
+          return(double_variable_get_values_at_index_vector(self$.variable, index))
+        }
       }
-      if (is.numeric(index)) {
-        stopifnot(all(index > 0))
-        return(double_variable_get_values_at_index_vector(self$.variable, index))
-      }
-      double_variable_get_values_at_index(self$.variable, index$.bitset)
     },
 
     #' @description return a \code{\link[individual]{Bitset}} giving individuals 
@@ -35,7 +38,7 @@ DoubleVariable <- R6Class(
     #' @param b upper bound
     get_index_of = function(a, b) {
       stopifnot(a < b)
-      return(Bitset$new(from = double_variable_get_index_of_range(self$.variable, a, b)))            
+      return(Bitset$new(from = double_variable_get_index_of_range(self$.variable, a, b)))      
     },
 
     #' @description return the number of individuals whose value lies in an interval
@@ -44,7 +47,7 @@ DoubleVariable <- R6Class(
     #' @param b upper bound
     get_size_of = function(a, b) {
       stopifnot(a < b)
-      double_variable_get_size_of_range(self$.variable, a, b)            
+      return(double_variable_get_size_of_range(self$.variable, a, b))
     },
 
     #' @description Queue an update for a variable. There are 4 types of variable update:
@@ -92,6 +95,7 @@ DoubleVariable <- R6Class(
           }
         } else {
           if (length(index) != 0) {
+            stopifnot(all(is.finite(index)))
             stopifnot(all(index > 0))
             double_variable_queue_update(
               self$.variable,